Initial commit
importing dependences and started to investigate the sensibility by
applying spearman's rank order correlation

+++ generate_emap.py
@@ -251,7 +251,7 @@ def clip_encode_dense(x):
     clip_ksize = clipmodel.visual.conv1.kernel_size
     
     # modified from CLIP
-    x = x.half()
+    #x = x.half()
     x = clipmodel.visual.conv1(x)  
     feah, feaw = x.shape[-2:]
